Update to current Cycle and Onionify
This change keeps up with various progress on core Cycle and especially the Onionify addon. Unfortunately the change is significant enough that it makes parts of the video series of it obsolete. I will consider recording a follow-up video describing these changes.

This change relies on an update to the cyclejs-modal library, currently in PR:

https://github.com/cyclejs-community/cyclejs-modal/pull/7